backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
torch
torchvision
pillow
python-multipart
requests
python-dotenv